ml_server/requirements.txt
Flask
torch
torchvision
Pillow